rename: hooks 및 테스트 파일 경로 변경
