+++ hooks/src/lib/index.ts
@@ -1,6 +1,6 @@
-export { default as useCardCompany } from "./hooks/useCardCompany";
-export { default as useCardHolder } from "./hooks/useCardHolder";
-export { default as useCardNumbers } from "./hooks/useCardNumbers";
-export { default as useCVCNumber } from "./hooks/useCVCNumber";
-export { default as useExpiryDate } from "./hooks/useExpiryDate";
-export { default as usePassword } from "./hooks/usePassword";
+export { default as useCardCompany } from "./hooks/useCardCompany/useCardCompany";
+export { default as useCardHolder } from "./hooks/useCardHolder/useCardHolder";
+export { default as useCardNumbers } from "./hooks/useCardNumbers/useCardNumbers";
+export { default as useCVCNumber } from "./hooks/useCVCNumber/useCVCNumber";
+export { default as useExpiryDate } from "./hooks/useExpiryDate/useExpiryDate";
+export { default as usePassword } from "./hooks/usePassword/usePassword";
